feat: 딜러가 블랙잭일 경우 게임 종료 로직 구현

+++ src/main/kotlin/blackjack/model/BettingMoney.kt
@@ -2,10 +2,10 @@ package blackjack.model
 
 class BettingMoney(
     value: Int,
-) : Money(value) {
+) : Money(value.toDouble()) {
     init {
         require(value in 1..300_000_000) { "베팅 금액은 최소 1원 이상 최대 3억원까지 가능합니다. $value 는 범위에 해당되지 않습니다." }
     }
 
-    fun getProfit(): Money = Money(value - initialValue)
+    fun getProfit(): Money = Money(value.toInt() - initialValue)
 }
